# Use snprintf() of standard C library