@Watson1978 2022-01-10
Here is stacktrace with before changing.
<img width="956" alt="スクリーンショット 2022-01-10 18 22 31" src="https://user-images.githubusercontent.com/199156/148743121-33b7f97f-c07f-446d-9769-30060c6dec39.png">